Moved a bunch of tests to their correct locations
