+++ obspy/io/ascii/tests/test_ascii.py
