+++ obspy/io/json/tests/__init__.py
@@ -8,7 +8,7 @@ import unittest
 from obspy.core.util import add_doctests, add_unittests
 
 
-MODULE_NAME = "obspy.core"
+MODULE_NAME = "obspy.io.json"
 
 
 def suite():
