+++ obspy/io/json/tests/test_json.py
