+++ obspy/scripts/tests/test_print.py
